rag/requirements.txt
faiss-cpu==1.8.0.post1
rank-bm25==0.2.2
sentence-transformers==2.7.0
